requirements.txt
asn1crypto==0.24.0
azure-common==1.1.8
azure-nspkg==2.0.0
azure-storage-blob==1.1.0
azure-storage-common==1.1.0
azure-storage-nspkg==3.0.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
ckan==git+https://github.com/open-data/ckan.git@canada-v2.6
ckanapi==4.1
configparser==3.5.0
cryptography==2.1.4
docopt==0.6.2
enum34==1.1.6
futures==3.2.0
idna==2.6
ipaddress==1.0.19
lxml==4.2.0
pycparser==2.18
python-dateutil==2.7.0
requests==2.18.4
simplejson==3.13.2
six==1.11.0
urllib3==1.22
uuid==1.30
pyyaml==3.12